Added a new approach to selecting tac ids for subscribers
The previous commit (c875a13cf9) was incorrect - as the subscriber count
increased, the modulus on the offset stopped working to set a reasonable
offset. This approach divides the brands into a cycling SEQUENCE that
will revolve the offset within a brand parameter to ensure a good spread
of tac ids.