Altered sqlalchemy engine "begin" to "connect"
When writing large data sets, because begin starts a transaction, it is
not possible to connect to the DB for the duration of the seed process
(which could be days). In changing this, it is possible to connect to
the DB and to query tables that aren't being written to.