Make airflow a peer dependency
Technically, airflow is required to run any of flowetl. But, we avoid at any point importing it outside of a functions because doing so writes to the user's home. It also pins specific versions of a large number of dependencies, which makes locking a nightmare. Hence, we make it an extra here and can then lock easily in docs build.